python/arabic/requirements.txt
torch>=1.9.0,<2.0.0
numpy>=1.20.0,<2.0.0
matplotlib>=3.3.3
pandas>=1.3.0
ruamel.yaml>=0.16.12
tensorboard>=2.4.0
diacritization-evaluation==0.5
tqdm>=4.56.0
onnx>=1.9.0
onnxruntime>=1.8.1
pyyaml>=5.4.1
